Fix index out of range error
The length of the variable containing the size of the board has to be
incremented, in order to prevent this error for larger boards (19 and
on).

It used to be a ``uint8``, which fits well for 19, but not for 19 x 19 =
361, because in this case was truncating to 105. Therefore when trying
to index, it was failing because the array was of length 105, and for
products greater than that, it was out of the limits.

The solution was to change the type for a ``uint16``, that for a board
size of 19, works.

Fixes https://github.com/ami-go/baduk-engine/issues/2

+++ abstractboard.go
@@ -14,18 +14,18 @@ type AbstractBoard struct {
 }
 
 // Creates new Go Board
-func NewBoard(boardSize uint8) (*AbstractBoard, error) {
+func NewBoard(boardSize uint16) (*AbstractBoard, error) {
 	if boardSize < 1 {
 		return nil, fmt.Errorf("Boardsize can not be less than 1!")
 	}
 
 	board := make([][]BoardStatus, boardSize)
-	for i := uint8(0); i < boardSize; i++ {
+	for i := uint8(0); i < uint8(boardSize); i++ {
 		board[i] = make([]BoardStatus, boardSize)
 	}
 
 	return &AbstractBoard{
-		boardSize,
+		uint8(boardSize),
 		board,
 		make([]*Move, 0),
 		NewZobristHash(boardSize),
@@ -293,11 +293,9 @@ func (board *AbstractBoard) getNeighbours(x, y uint8) (neighbourIndexes []Positi
 }
 
 func (board *AbstractBoard) getStatus(x, y uint8) BoardStatus {
-
 	return board.data[x][y]
 }
 
 func (board *AbstractBoard) setStatus(x uint8, y uint8, status BoardStatus) {
-
 	board.data[x][y] = status
 }
